+++ src/punctuators.cpp
@@ -113,7 +113,7 @@ static const chunk_tag_t symbols1[] =
 
 #include "punctuators.h"
 
-const chunk_tag_t *find_punctuator(const char *str, UINT8 lang_flags)
+const chunk_tag_t *find_punctuator(const char *str, int lang_flags)
 {
    const chunk_tag_t    *p_match = NULL;
    const lookup_entry_t *p_tab   = punc_table;
